fixup! perf_hooks: reduce overhead of createHistogram
